requirements.txt
datasets==3.2.0
pandas==2.3.3
peft==0.14.0
Pillow==12.0.0
torch==2.5.1+cu118
transformers==4.48.3
